cleanup 2 - squash required

+++ src/test/bridges/convex/ConvexStakingBridge.t.sol
@@ -486,7 +486,7 @@ contract ConvexStakingBridgeTest is BridgeTestBase {
 
     function _setupInvalidPoolIds() internal {
         for (uint256 pid = 0; pid < invalidPids.length; pid++) {
-            invalidPoolIds[uint16(invalidPids[pid])] = true;
+            invalidPoolIds[invalidPids[pid]] = true;
         }
     }
 }
